news_shorts/requirements.txt
openai==1.59.3
moviepy==1.0.3
edge-tts==6.1.13
requests==2.32.3
python-dotenv==1.0.1
pillow==11.1.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
numpy<2.0.0
playwright
nest_asyncio
feedparser
schedule
beautifulsoup4
dateparser
